Move line drawing stuff into Line.{cpp.h}
