+++ src/openrct2/drawing/Line.cpp
@@ -7,164 +7,213 @@
  * OpenRCT2 is licensed under the GNU General Public License version 3.
  *****************************************************************************/
 
+#include "Line.h"
+
+#include "../interface/ScreenCoords.hpp"
 #include "Drawing.h"
+#include "IDrawingContext.h"
+#include "IDrawingEngine.h"
 #include "RenderTarget.h"
 
 #include <cmath>
 
-using namespace OpenRCT2::Drawing;
-
-/**
- * Draws a horizontal line of specified colour to a buffer.
- *  rct2: 0x0068474C
- */
-static void GfxDrawLineOnBuffer(RenderTarget& rt, PaletteIndex colour, const ScreenCoordsXY& coords, int32_t no_pixels)
+namespace OpenRCT2::Drawing
 {
-    ScreenCoordsXY offset{ coords.x - rt.x, coords.y - rt.y };
-
-    const int32_t width = rt.width;
-    const int32_t height = rt.height;
-
-    // Check to make sure point is in the y range
-    if (offset.y < 0)
-        return;
-    if (offset.y >= height)
-        return;
-    // Check to make sure we are drawing at least a pixel
-    if (!no_pixels)
-        no_pixels++;
-
-    // If x coord outside range leave
-    if (offset.x < 0)
+    /**
+     * Draws a horizontal line of specified colour to a buffer.
+     *  rct2: 0x0068474C
+     */
+    static void GfxDrawLineOnBuffer(RenderTarget& rt, PaletteIndex colour, const ScreenCoordsXY& coords, int32_t no_pixels)
     {
-        // Unless the number of pixels is enough to be in range
-        no_pixels += offset.x;
-        if (no_pixels <= 0)
+        ScreenCoordsXY offset{ coords.x - rt.x, coords.y - rt.y };
+
+        const int32_t width = rt.width;
+        const int32_t height = rt.height;
+
+        // Check to make sure point is in the y range
+        if (offset.y < 0)
             return;
-        // Resets starting point to 0 as we don't draw outside the range
-        offset.x = 0;
-    }
-
-    // Ensure that the end point of the line is within range
-    if (offset.x + no_pixels - width > 0)
-    {
-        // If the end point has any pixels outside range
-        // cut them off. If there are now no pixels return.
-        no_pixels -= offset.x + no_pixels - width;
-        if (no_pixels <= 0)
+        if (offset.y >= height)
             return;
+        // Check to make sure we are drawing at least a pixel
+        if (!no_pixels)
+            no_pixels++;
+
+        // If x coord outside range leave
+        if (offset.x < 0)
+        {
+            // Unless the number of pixels is enough to be in range
+            no_pixels += offset.x;
+            if (no_pixels <= 0)
+                return;
+            // Resets starting point to 0 as we don't draw outside the range
+            offset.x = 0;
+        }
+
+        // Ensure that the end point of the line is within range
+        if (offset.x + no_pixels - width > 0)
+        {
+            // If the end point has any pixels outside range
+            // cut them off. If there are now no pixels return.
+            no_pixels -= offset.x + no_pixels - width;
+            if (no_pixels <= 0)
+                return;
+        }
+
+        // Get the buffer we are drawing to and move to the first coordinate.
+        PaletteIndex* bits_pointer = reinterpret_cast<PaletteIndex*>(rt.bits + offset.y * rt.LineStride() + offset.x);
+
+        // Draw the line to the specified colour
+        for (; no_pixels > 0; --no_pixels, ++bits_pointer)
+        {
+            *bits_pointer = colour;
+        }
     }
 
-    // Get the buffer we are drawing to and move to the first coordinate.
-    PaletteIndex* bits_pointer = reinterpret_cast<PaletteIndex*>(rt.bits + offset.y * rt.LineStride() + offset.x);
+    /**
+     * Draws a line on rt if within rt boundaries
+     *  rct2: 0x00684466
+     * rt (edi)
+     * x1 (ax)
+     * y1 (bx)
+     * x2 (cx)
+     * y2 (dx)
+     * colour (ebp)
+     */
 
-    // Draw the line to the specified colour
-    for (; no_pixels > 0; --no_pixels, ++bits_pointer)
+    void GfxDrawLineSoftware(RenderTarget& rt, const ScreenLine& line, PaletteIndex colour)
     {
-        *bits_pointer = colour;
-    }
-}
+        const ZoomLevel zoom = rt.zoom_level;
+        int32_t x1 = zoom.ApplyInversedTo(line.GetX1());
+        int32_t x2 = zoom.ApplyInversedTo(line.GetX2());
+        int32_t y1 = zoom.ApplyInversedTo(line.GetY1());
+        int32_t y2 = zoom.ApplyInversedTo(line.GetY2());
+        // Check to make sure the line is within the drawing area
+        if ((x1 < rt.x) && (x2 < rt.x))
+        {
+            return;
+        }
 
-/**
- * Draws a line on rt if within rt boundaries
- *  rct2: 0x00684466
- * rt (edi)
- * x1 (ax)
- * y1 (bx)
- * x2 (cx)
- * y2 (dx)
- * colour (ebp)
- */
+        if ((y1 < rt.y) && (y2 < rt.y))
+        {
+            return;
+        }
 
-void GfxDrawLineSoftware(RenderTarget& rt, const ScreenLine& line, PaletteIndex colour)
-{
-    const ZoomLevel zoom = rt.zoom_level;
-    int32_t x1 = zoom.ApplyInversedTo(line.GetX1());
-    int32_t x2 = zoom.ApplyInversedTo(line.GetX2());
-    int32_t y1 = zoom.ApplyInversedTo(line.GetY1());
-    int32_t y2 = zoom.ApplyInversedTo(line.GetY2());
-    // Check to make sure the line is within the drawing area
-    if ((x1 < rt.x) && (x2 < rt.x))
-    {
-        return;
-    }
+        if ((x1 > (rt.x + rt.width)) && (x2 > (rt.x + rt.width)))
+        {
+            return;
+        }
 
-    if ((y1 < rt.y) && (y2 < rt.y))
-    {
-        return;
-    }
+        if ((y1 > (rt.y + rt.height)) && (y2 > (rt.y + rt.height)))
+        {
+            return;
+        }
 
-    if ((x1 > (rt.x + rt.width)) && (x2 > (rt.x + rt.width)))
-    {
-        return;
-    }
+        // Bresenham's algorithm
 
-    if ((y1 > (rt.y + rt.height)) && (y2 > (rt.y + rt.height)))
-    {
-        return;
-    }
-
-    // Bresenham's algorithm
-
-    // If vertical plot points upwards
-    int32_t steep = std::abs(y2 - y1) > std::abs(x2 - x1);
-    if (steep)
-    {
-        int32_t temp_y2 = y2;
-        int32_t temp_x2 = x2;
-        y2 = x1;
-        x2 = y1;
-        y1 = temp_x2;
-        x1 = temp_y2;
-    }
-
-    // If line is right to left swap direction
-    if (x1 > x2)
-    {
-        int32_t temp_y2 = y2;
-        int32_t temp_x2 = x2;
-        y2 = y1;
-        x2 = x1;
-        y1 = temp_y2;
-        x1 = temp_x2;
-    }
-
-    int32_t delta_x = x2 - x1;
-    int32_t delta_y = std::abs(y2 - y1);
-    int32_t error = delta_x / 2;
-    int32_t y_step;
-    int32_t y = y1;
-
-    // Direction of step
-    if (y1 < y2)
-        y_step = 1;
-    else
-        y_step = -1;
-
-    for (int32_t x = x1, x_start = x1, length = 1; x < x2; ++x, ++length)
-    {
-        // Vertical lines are drawn 1 pixel at a time
+        // If vertical plot points upwards
+        int32_t steep = std::abs(y2 - y1) > std::abs(x2 - x1);
         if (steep)
-            GfxDrawLineOnBuffer(rt, colour, { y, x }, 1);
-
-        error -= delta_y;
-        if (error < 0)
         {
-            // Non vertical lines are drawn with as many pixels in a horizontal line as possible
-            if (!steep)
-                GfxDrawLineOnBuffer(rt, colour, { x_start, y }, length);
-
-            // Reset non vertical line vars
-            x_start = x + 1;
-            length = 0; // NB: will be incremented in next iteration
-            y += y_step;
-            error += delta_x;
+            int32_t temp_y2 = y2;
+            int32_t temp_x2 = x2;
+            y2 = x1;
+            x2 = y1;
+            y1 = temp_x2;
+            x1 = temp_y2;
         }
 
-        // Catch the case of the last line
-        if (x + 1 == x2 && !steep)
+        // If line is right to left swap direction
+        if (x1 > x2)
         {
-            GfxDrawLineOnBuffer(rt, colour, { x_start, y }, length);
+            int32_t temp_y2 = y2;
+            int32_t temp_x2 = x2;
+            y2 = y1;
+            x2 = x1;
+            y1 = temp_y2;
+            x1 = temp_x2;
+        }
+
+        int32_t delta_x = x2 - x1;
+        int32_t delta_y = std::abs(y2 - y1);
+        int32_t error = delta_x / 2;
+        int32_t y_step;
+        int32_t y = y1;
+
+        // Direction of step
+        if (y1 < y2)
+            y_step = 1;
+        else
+            y_step = -1;
+
+        for (int32_t x = x1, x_start = x1, length = 1; x < x2; ++x, ++length)
+        {
+            // Vertical lines are drawn 1 pixel at a time
+            if (steep)
+                GfxDrawLineOnBuffer(rt, colour, { y, x }, 1);
+
+            error -= delta_y;
+            if (error < 0)
+            {
+                // Non vertical lines are drawn with as many pixels in a horizontal line as possible
+                if (!steep)
+                    GfxDrawLineOnBuffer(rt, colour, { x_start, y }, length);
+
+                // Reset non vertical line vars
+                x_start = x + 1;
+                length = 0; // NB: will be incremented in next iteration
+                y += y_step;
+                error += delta_x;
+            }
+
+            // Catch the case of the last line
+            if (x + 1 == x2 && !steep)
+            {
+                GfxDrawLineOnBuffer(rt, colour, { x_start, y }, length);
+            }
         }
     }
-}
+
+    void GfxDrawLine(RenderTarget& rt, const ScreenLine& line, PaletteIndex colour)
+    {
+        auto drawingEngine = rt.DrawingEngine;
+        if (drawingEngine != nullptr)
+        {
+            IDrawingContext* dc = drawingEngine->GetDrawingContext();
+            dc->DrawLine(rt, colour, line);
+        }
+    }
+
+    void GfxDrawDashedLine(
+        RenderTarget& rt, const ScreenLine& screenLine, const int32_t dashedLineSegmentLength, const PaletteIndex colour)
+    {
+        assert(dashedLineSegmentLength > 0);
+
+        const auto drawingEngine = rt.DrawingEngine;
+        if (drawingEngine != nullptr)
+        {
+            constexpr int32_t kPrecisionFactor = 1000;
+
+            const int32_t dashedLineLength = std::hypot(
+                screenLine.GetX2() - screenLine.GetX1(), screenLine.GetY2() - screenLine.GetY1());
+            const int32_t lineSegmentCount = dashedLineLength / dashedLineSegmentLength / 2;
+            if (lineSegmentCount == 0)
+            {
+                return;
+            }
+
+            const int32_t lineXDist = std::abs(screenLine.GetX2() - screenLine.GetX1());
+            const int32_t lineYDist = std::abs(screenLine.GetY2() - screenLine.GetY1());
+            const int32_t dxPrecise = kPrecisionFactor * lineXDist / lineSegmentCount / 2;
+            const int32_t dyPrecise = kPrecisionFactor * lineYDist / lineSegmentCount / 2;
+            IDrawingContext* dc = drawingEngine->GetDrawingContext();
+
+            for (int32_t i = 0, x, y; i < lineSegmentCount; ++i)
+            {
+                x = screenLine.GetX1() + dxPrecise * i * 2 / kPrecisionFactor;
+                y = screenLine.GetY1() + dyPrecise * i * 2 / kPrecisionFactor;
+                dc->DrawLine(rt, colour, { { x, y }, { x + dxPrecise / kPrecisionFactor, y + dyPrecise / kPrecisionFactor } });
+            }
+        }
+    }
+} // namespace OpenRCT2::Drawing
