+++ src/openrct2/libopenrct2.vcxproj
@@ -265,6 +265,7 @@
     <ClInclude Include="drawing\ImageImporter.h" />
     <ClInclude Include="drawing\InvalidationGrid.h" />
     <ClInclude Include="drawing\LightFX.h" />
+    <ClInclude Include="drawing\Line.h" />
     <ClInclude Include="drawing\NewDrawing.h" />
     <ClInclude Include="drawing\PaletteIndex.h" />
     <ClInclude Include="drawing\Palette.h" />
